refactor: split types for watch data and nvapi response
use typescript references with eslint

resolve package root path modules
