+++ package.json
@@ -37,6 +37,7 @@
     "@typescript-eslint/eslint-plugin": "^6.19.1",
     "@typescript-eslint/parser": "^6.19.1",
     "@violentmonkey/types": "^0.1.7",
+    "babel-plugin-tsconfig-paths": "^1.0.3",
     "colorette": "^2.0.20",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
