+++ packages/build-util/package.json
@@ -6,5 +6,6 @@
   "scripts": {},
   "author": "kPherox <kphrx@kpherox.dev>",
   "license": "WTFPL",
-  "type": "module"
+  "type": "module",
+  "types": "dist/index.d.ts"
 }
