+++ packages/script/zenza-watch/src/components/hover-menu/mouse-controller.ts
@@ -1,4 +1,5 @@
 import type {ReactiveController} from "lit";
+
 import type {LeftHoverMenu, RightHoverMenu} from "./index";
 
 type ReactiveControllerHost = LeftHoverMenu | RightHoverMenu;
